backend/pyproject.toml
[project]
name = "karamba"
version = "0.1.0"
description = "Reusable AI Agent Framework with Document Understanding"
authors = [{name = "Jose Santiago", email = "jfsantiagonunez@gmail.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    
    # LLM & AI
    "litellm>=1.30.0",
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    "langgraph>=0.2.60",
    "langgraph-checkpoint>=2.0.9",
    "langgraph-checkpoint-sqlite>=2.0.5",
    "langchain-core>=0.3.29",
    
    # Document Processing
    "pypdfium2>=4.26.0",
    "python-docx>=1.1.0",
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    
    # Vector Search & Embeddings
    "sentence-transformers>=2.3.0",
    "chromadb>=0.4.22",
    "faiss-cpu>=1.7.4",
    
    # Utilities
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.98.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true